2.18: fixed compiling bugs, converted things into combat rolls, removed non-existing references, ported some things to 64-bit etc...

+++ core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/actors/blobs/ToxicGas.java
@@ -40,7 +40,7 @@ public class ToxicGas extends Blob implements Hero.Doom {
 	protected void evolve() {
 		super.evolve();
 
-		int damage = 1 + Dungeon.scalingDepth()/5;
+		long damage = 1 + Dungeon.scalingDepth()/5;
 
 		Char ch;
 		int cell;
