+++ core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/items/trinkets/MimicTooth.java
@@ -33,14 +33,14 @@ public class MimicTooth extends Trinket {
 	@Override
 	protected int upgradeEnergyCost() {
 		//5 -> 8(13) -> 10(23) -> 12(35)
-		return 6+2*level();
+		return 6+2*(int)level();
 	}
 
 	@Override
 	public String desc() {
 		return Messages.get(this, "desc",
-				Messages.decimalFormat("#.##", mimicChanceMultiplier(buffedLvl())),
-				Messages.decimalFormat("#.##", 100*ebonyMimicChance(buffedLvl())));
+				Messages.decimalFormat("#.##", mimicChanceMultiplier((int)buffedLvl())),
+				Messages.decimalFormat("#.##", 100*ebonyMimicChance((int)buffedLvl())));
 	}
 
 	public static float mimicChanceMultiplier(){
