+++ core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/items/trinkets/RatSkull.java
@@ -33,12 +33,12 @@ public class RatSkull extends Trinket {
 	@Override
 	protected int upgradeEnergyCost() {
 		//5 -> 8(13) -> 10(23) -> 12(35)
-		return 6+2*level();
+		return 6+2*(int)level();
 	}
 
 	@Override
 	public String desc() {
-		return Messages.get(this, "desc", (int)(exoticChanceMultiplier(buffedLvl())));
+		return Messages.get(this, "desc", (int)(exoticChanceMultiplier((int)buffedLvl())));
 	}
 
 	public static float exoticChanceMultiplier(){
