+++ core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/items/trinkets/TrapMechanism.java
@@ -39,12 +39,12 @@ public class TrapMechanism extends Trinket {
 	@Override
 	protected int upgradeEnergyCost() {
 		//5 -> 8(13) -> 10(23) -> 12(35)
-		return 6+2*level();
+		return 6+2*(int)level();
 	}
 
 	@Override
 	public String desc() {
-		return Messages.get(this, "desc", (int)(100*overrideNormalLevelChance(buffedLvl())));
+		return Messages.get(this, "desc", (int)(100*overrideNormalLevelChance((int)buffedLvl())));
 	}
 
 	public static float overrideNormalLevelChance(){
